2018-8-10提交

+++ src/main/java/com/zltel/broadcast/resource/service/impl/MaterialAlbumServiceImpl.java
@@ -13,8 +13,9 @@ import com.zltel.broadcast.common.support.BaseDao;
 import com.zltel.broadcast.common.support.BaseDaoImpl;
 import com.zltel.broadcast.resource.bean.Material;
 import com.zltel.broadcast.resource.bean.MaterialAlbum;
+import com.zltel.broadcast.resource.bean.ResourceMaterial;
 import com.zltel.broadcast.resource.dao.MaterialAlbumMapper;
-import com.zltel.broadcast.resource.dao.MaterialMapper;
+import com.zltel.broadcast.resource.dao.ResourceMaterialMapper;
 import com.zltel.broadcast.resource.service.MaterialAlbumService;
 
 @Service
@@ -23,7 +24,7 @@ public class MaterialAlbumServiceImpl extends BaseDaoImpl<MaterialAlbum> impleme
     private MaterialAlbumMapper mapper;
 
     @Resource
-    private MaterialMapper materialMapper;
+    private ResourceMaterialMapper resourceMaterialMapper;
 
     @Override
     public BaseDao<MaterialAlbum> getInstince() {
@@ -33,9 +34,13 @@ public class MaterialAlbumServiceImpl extends BaseDaoImpl<MaterialAlbum> impleme
     @Override
     public int delete(MaterialAlbum ma) {
         PageRowBounds prb = new PageRowBounds(0, 0);
-        Material r = new Material();
+        ResourceMaterial r = new ResourceMaterial();
         r.setAlbumId(ma.getAlbumId());
-        this.materialMapper.query(r, prb);
+        MaterialAlbum dt = this.mapper.selectByPrimaryKey(ma.getAlbumId());
+        if (dt == null) throw new RRException("没有找到要删除数据");
+        if (dt.getBuiltin()) throw new RRException("无法删除内置节点");
+        
+        this.resourceMaterialMapper.query(r, prb);
         if (prb.getTotal() > 0) throw new RRException("此分类下包含数据,无法删除!");
 
         return this.mapper.deleteByPrimaryKey(ma.getAlbumId());
